feat: add image upload functionality and pagination
- Added multiple image files for product management.
- Updated package dependencies: added multer for file uploads and pdfkit for PDF generation.
- Enhanced form validation for product title length.
- Implemented pagination in product listing and order views.
- Created a new utility function for file deletion.
- Added a 500 error page for better error handling.
- Updated product forms to support image uploads instead of URLs.
- Fixed image source paths in product detail and list views.
- Added invoice link to orders for better user experience.

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const Product = require("../models/product");
 const { validationResult } = require("express-validator");
-
+const path = require("../util/path");
+const fileHelper = require("../util/file");
 exports.getAddProduct = (req, res, next) => {
   if (!req.session.isLoggedIn) {
     return res.redirect("/login");
@@ -16,9 +17,21 @@ exports.getAddProduct = (req, res, next) => {
 };
 
 exports.postAddProduct = async (req, res, next) => {
-  const { title, price, description, imageUrl } = req.body;
+  const { title, price, description } = req.body;
+  const image = req.file;
+
+  if (!image) {
+    return res.status(422).render("admin/edit-product", {
+      pageTitle: "New Product",
+      path: "/admin/add-product",
+      editing: false,
+      errorMessage: [{ msg: "Attached file is not an image" }],
+      oldInput: { title, price, description, image },
+    });
+  }
+
+  const imageUrl = image.path;
   const errors = validationResult(req);
-  console.log({ errors: errors.array() });
 
   if (!errors.isEmpty()) {
     return res.status(422).render("admin/edit-product", {
@@ -30,14 +43,21 @@ exports.postAddProduct = async (req, res, next) => {
     });
   }
 
-  const prod = new Product({
-    title,
-    price,
-    description,
-    imageUrl,
-    userId: req.user,
-  });
-  await prod.save();
+  try {
+    const prod = new Product({
+      title,
+      price,
+      description,
+      imageUrl,
+      userId: req.user,
+    });
+    await prod.save();
+  } catch (error) {
+    console.log(error);
+    const err = new Error(error);
+    err.httpStatusCode = 500;
+    return next(err);
+  }
   res.redirect("/");
 };
 
@@ -51,44 +71,62 @@ exports.getEditProduct = async (req, res, next) => {
     return res.redirect("/");
   }
 
-  const product = await Product.findById(productId);
-  if (!product) {
-    return res.redirect("/");
+  try {
+    const product = await Product.findById(productId);
+    if (!product) {
+      return res.redirect("/");
+    }
+    res.render("admin/edit-product", {
+      pageTitle: "New Product",
+      path: "/admin/add-product",
+      editing: edit,
+      oldInput: product,
+      errorMessage: req.flash("error"),
+    });
+  } catch (error) {
+    const err = new Error(error);
+    err.httpStatusCode = 500;
+    return next(err);
   }
-  res.render("admin/edit-product", {
-    pageTitle: "New Product",
-    path: "/admin/add-product",
-    editing: edit,
-    oldInput: product,
-    errorMessage: req.flash("error"),
-  });
 };
 
 exports.postEditProducts = async (req, res, next) => {
-  const { title, price, description, imageUrl } = req.body;
+  const { title, price, description } = req.body;
+  const image = req.file;
   const id = req.params.productId;
-  const product = await Product.findById(id);
-  const errors = validationResult(req);
-  console.log({ errors: errors.array() });
+  try {
+    const product = await Product.findById(id);
+    const errors = validationResult(req);
 
-  if (!errors.isEmpty()) {
-    return res.status(422).render("admin/edit-product", {
-      pageTitle: "New Product",
-      path: "/admin/add-product",
-      editing: true,
-      errorMessage: errors.array(),
-      oldInput: { _id: id, title, price, description, imageUrl },
-    });
-  }
-
-  if (product) {
-    if (product.userId.toString() !== req.user._id.toString()) {
-      return res.redirect("/");
+    if (!errors.isEmpty()) {
+      return res.status(422).render("admin/edit-product", {
+        pageTitle: "New Product",
+        path: "/admin/add-product",
+        editing: true,
+        errorMessage: errors.array(),
+        oldInput: { _id: id, title, price, description, image: undefined },
+      });
     }
-    await product.updateOne({ title, price, description, imageUrl });
-  }
 
-  res.redirect("/admin/products");
+    if (product) {
+      let imageUrl = product.imageUrl;
+      if (image?.path) {
+        fileHelper.deleteFile(imageUrl);
+        imageUrl = image.path;
+      }
+
+      if (product.userId.toString() !== req.user._id.toString()) {
+        return res.redirect("/");
+      }
+      await product.updateOne({ title, price, description, imageUrl });
+    }
+
+    res.redirect("/admin/products");
+  } catch (error) {
+    const err = new Error(error);
+    err.httpStatusCode = 500;
+    return next(err);
+  }
 };
 
 exports.getProducts = async (req, res, next) => {
@@ -104,10 +142,23 @@ exports.getProducts = async (req, res, next) => {
 
 exports.deleteProduct = async (req, res, next) => {
   const productId = req.params.productId;
-  await Product.deleteOne({
-    _id: productId,
-    userId: req.user._id,
-  });
+
+  try {
+    const existProd = await Product.findById(productId);
+    if (!existProd) {
+      return next(new Error("Product not found"));
+    }
+
+    fileHelper.deleteFile(existProd.imageUrl)
+    await Product.deleteOne({
+      _id: productId,
+      userId: req.user._id,
+    });
+  } catch (error) {
+    const err = new Error(error);
+    err.httpStatusCode = 500;
+    return next(err);
+  }
 
   res.redirect("/admin/products");
 };
